extracted compiler code
